Split connection code into submodules.
