+++ lib/ansible/runner/connection/paramiko_ssh.py
@@ -36,32 +36,6 @@ with warnings.catch_warnings():
     warnings.simplefilter("ignore")
     import paramiko
 
-
-################################################
-
-class Connection(object):
-    ''' Handles abstract connections to remote hosts '''
-
-    def __init__(self, runner, transport,sudo_user):
-        self.runner = runner
-        self.transport = transport
-        self.sudo_user = sudo_user
-    def connect(self, host, port=None):
-        conn = None
-        if self.transport == 'local':
-            conn = LocalConnection(self.runner, host)
-        elif self.transport == 'paramiko':
-            conn = ParamikoConnection(self.runner, host, port)
-        if conn is None:
-            raise Exception("unsupported connection type")
-        return conn.connect()
-
-################################################
-# want to implement another connection type?
-# follow duck-typing of ParamikoConnection
-# you may wish to read config files in __init__
-# if you have any.  Paramiko does not need any.
-
 class ParamikoConnection(object):
     ''' SSH based connections with Paramiko '''
 
@@ -171,54 +145,3 @@ class ParamikoConnection(object):
 
         self.ssh.close()
 
-############################################
-# add other connection types here
-
-class LocalConnection(object):
-    ''' Local based connections '''
-
-    def __init__(self, runner, host):
-        self.runner = runner
-        self.host = host
-
-    def connect(self, port=None):
-        ''' connect to the local host; nothing to do here '''
-
-        return self
-
-    def exec_command(self, cmd, tmp_path,sudo_user,sudoable=False):
-        ''' run a command on the local host '''
-        if self.runner.sudo and sudoable:
-            cmd = "sudo -s %s" % cmd
-        if self.runner.sudo_pass:
-            # NOTE: if someone wants to add sudo w/ password to the local connection type, they are welcome
-            # to do so.  The primary usage of the local connection is for crontab and kickstart usage however
-            # so this doesn't seem to be a huge priority
-            raise errors.AnsibleError("sudo with password is presently only supported on the paramiko (SSH) connection type")
-
-        p = subprocess.Popen(cmd, shell=True, stdin=None,
-                             stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-        stdout, stderr = p.communicate()
-        return ("", stdout, stderr)
-
-    def put_file(self, in_path, out_path):
-        ''' transfer a file from local to local '''
-        if not os.path.exists(in_path):
-            raise errors.AnsibleFileNotFound("file or module does not exist: %s" % in_path)
-        try:
-            shutil.copyfile(in_path, out_path)
-        except shutil.Error:
-            traceback.print_exc()
-            raise errors.AnsibleError("failed to copy: %s and %s are the same" % (in_path, out_path))
-        except IOError:
-            traceback.print_exc()
-            raise errors.AnsibleError("failed to transfer file to %s" % out_path)
-
-    def fetch_file(self, in_path, out_path):
-        ''' fetch a file from local to local -- for copatibility '''
-        self.put_file(in_path, out_path)
-
-    def close(self):
-        ''' terminate the connection; nothing to do here '''
-
-        pass
